+++ docs/changes.rst
@@ -14,6 +14,7 @@ Version 0.5.7
 
 Unreleased.
 
+ - Added :meth:`Image.ordered_dither() <wand.image.BaseImage.ordered_dither>` method.
  - Added :meth:`Image.read_mask() <wand.image.BaseImage.read_mask>` method. [:issue:`433`]
  - Added :meth:`Image.sepia_tone() <wand.image.BaseImage.sepia_tone>` method.
  - Added :meth:`Image.swirl() <wand.image.BaseImage.swirl>` method.